Split the net commit orchestration by topic
Apply.cs was 2356 lines and 94 members: the per-frame commit cycle, temp-entity
isolation, the drain-and-recover path, three separate transaction validators,
owner resolution across two peers that share no entity ids, and the arming
entry points. Each of those is a subject; together they were a file nobody
reads top to bottom.

Split into ten partial files of 180 to 360 lines - temps, commit, drain, the
route/object/net validators, owner, topology, arm - each with a note saying
what belongs in it, and Apply.cs left holding the frame cycle and the state the
others read.

Pure code movement, verified member by member: the only textual change is the
blank line each file no longer needs before its first member.
